+++ releasenotes.md
@@ -5,5 +5,6 @@ Whats New in v10.1
 - Added orientation when exporting to a formula for selling
 - Allow ordering in formula sell table
 - Fix export functions in Formula Make page
-
+- Added export to JSON for perfume types
+- Added export to JSON for categories
 - For full details please refer to the CHANGELOG